move segment pointer outside texel check
The previous code was writing external segment pointers only if the material being written has texel data. It was also omitting tile size on that condition. External segments are not guaranteed to contain opcodes for tile size (for example: they may instead alter the color registers), and materials should be able to branch to external segments whether or not texel data is present. The code was slightly rearranged to address this.